requirements.txt
Django==1.10
django-multiselectfield==0.1.8
